update footer styles to fix carousels
