--- app/javascript/components/footer/footer-styles.scss
@@ -1,260 +1,0 @@
-@import '~styles/settings.scss';
-
-.c-footer {
-  font-family: $font-family-1;
-  background: #eceee8;
-
-  .footer-subscribe {
-    width: 100%;
-    height: 112px;
-    background-color: #fff;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-
-    .footer-subscribe-button {
-      width: 260px;
-      height: 40px;
-      background-color: #97bd3d;
-      color: #fff;
-      border-radius: 36px;
-      font-size: 14px;
-      cursor: pointer;
-    }
-  }
-
-  .footer-main {
-    padding: 0 20px;
-    position: relative;
-    margin: 0 auto;
-    width: 100%;
-    max-width: 1120px;
-
-    .footer-list {
-      display: block;
-      padding: 46px 0 25px;
-      border-bottom: 2px solid #d9d9d9;
-
-      .footer-links {
-        display: flex;
-        justify-content: space-between;
-        flex-direction: row;
-        align-items: center;
-      }
-
-      .footer-links-texts {
-        display: flex;
-        flex-flow: row wrap;
-
-        li {
-          flex-basis: auto;
-          margin: 0 45px 20px 0;
-        }
-      }
-
-      a {
-        font-size: 14px;
-        line-height: 1.86;
-        letter-spacing: 0.1px;
-        color: #666;
-        text-transform: uppercase;
-      }
-
-      .footer-links-social {
-        margin-top: 0;
-        display: flex;
-
-        li {
-          border: 1px solid rgba(0, 0, 153, 0.2);
-          width: 48px;
-          height: 48px;
-          border-radius: 100%;
-          font-size: 0;
-          margin-left: 17px;
-
-          a {
-            display: flex;
-            justify-content: center;
-            align-items: center;
-            width: 100%;
-            height: 100%;
-          }
-        }
-
-        .c-icon {
-          fill: #444;
-          height: 25px;
-          width: 25px;
-        }
-      }
-
-      .-links-contact {
-        display: flex;
-        flex-direction: row-reverse;
-        margin-top: 25px;
-        justify-content: space-between;
-        align-items: center;
-
-        .text-button {
-          margin-bottom: 0;
-          font-size: 12px;
-          letter-spacing: 0.1px;
-          color: #999;
-        }
-
-        .footer-links__contact {
-          position: relative;
-          padding-right: 5px;
-          cursor: pointer;
-        }
-      }
-
-      ul {
-        padding: 0;
-        list-style: none;
-      }
-    }
-
-    .footer-logos {
-      max-width: calc(100% - 30px);
-    }
-
-    .footer-info {
-      display: flex;
-      flex-direction: row;
-      justify-content: space-between;
-      align-items: center;
-      padding-top: 27px;
-
-      .footer-partner {
-        text-align: left;
-        margin: inherit;
-        align-self: flex-start;
-
-        p {
-          font-size: 14px;
-          letter-spacing: 0.1px;
-          color: #999;
-        }
-
-        .c-icon {
-          margin: 26px 0 0 -5px;
-          font-size: 18px;
-          line-height: 27px;
-          width: 143px;
-          height: 47px;
-          top: 20px;
-          position: relative;
-        }
-      }
-    }
-
-    .footer-terms {
-      text-align: right;
-      padding: 10px 0 24px;
-      color: #999;
-      font-size: 12px;
-      letter-spacing: 0.1px;
-      line-height: 18px;
-
-      .terms {
-        display: inline-block;
-        color: inherit;
-        text-decoration: none;
-      }
-
-      .terms:hover {
-        text-decoration: underline;
-      }
-    }
-  }
-
-  @media screen and (max-width: 850px) { //$screen-m
-    .footer-main {
-      padding: 0 15px;
-
-      .footer-list .footer-links {
-        display: flex;
-        flex-direction: column;
-        align-items: baseline;
-
-        .footer-links-texts {
-          li {
-            flex-basis: 100%;
-            margin-bottom: 15px;
-          }
-        }
-
-        .footer-links-social {
-          margin-top: 0;
-
-          li {
-            margin-left: 0;
-            margin-right: 20px;
-          }
-        }
-      }
-
-      .footer-info {
-        display: flex;
-        flex-direction: column-reverse;
-
-        .footer-partner {
-          margin: 30px auto;
-          align-self: flex-start;
-          text-align: center;
-
-          .c-icon {
-            position: initial;
-          }
-        }
-      }
-
-      .footer-terms {
-        text-align: center;
-
-        a {
-          display: block;
-        }
-      }
-    }
-  }
-
-  .partners-slide {
-    max-width: calc(100% - 30px);
-    margin: 0;
-
-    @media screen and (min-width: $screen-m) {
-      margin: rem(15px) rem(15px) 0 0;
-      max-width: rem(730px);
-    }
-  }
-
-  .slick-list {
-    overflow: hidden;
-  }
-
-  .slick-slide {
-    align-items: center;
-    justify-content: center;
-
-    img {
-      font-size: 5px;
-      color: black;
-      text-decoration: none;
-      margin: auto;
-    }
-  }
-
-  .slick-arrow {
-    top: calc(50% - 20px);
-  }
-
-  .slick-next {
-    right: -30px;
-  }
-
-  .slick-prev {
-    left: -30px;
-  }
-}
