+++ app/javascript/components/footer/styles.scss
@@ -1,0 +1,182 @@
+@import '~styles/settings.scss';
+
+.c-footer {
+  font-family: $font-family-1;
+  background: #eceee8;
+
+  .footer-subscribe {
+    width: 100%;
+    height: rem(110px);
+    background-color: #fff;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+
+    .footer-subscribe-button {
+      width: rem(260px);
+    }
+  }
+
+  .footer-main {
+    .footer-links {
+      padding: rem(30px) 0;
+      border-bottom: solid 1px #aaa;
+
+      a,
+      button {
+        font-size: 14px;
+        line-height: 1.86;
+        letter-spacing: 0.1px;
+        color: $slate;
+        text-transform: uppercase;
+        cursor: pointer;
+
+        &:hover {
+          color: darken($slate, 20%);
+        }
+      }
+
+      svg {
+        fill: #444;
+        height: 25px;
+        width: 25px;
+      }
+
+      .footer-links-texts {
+        height: 100%;
+        display: flex;
+        align-items: center;
+
+        li {
+          margin-right: rem(25px);
+        }
+      }
+
+      .footer-links-social {
+        display: flex;
+        margin-top: rem(15px);
+
+        @media screen and (min-width: $screen-m) {
+          justify-content: flex-end;
+          margin-top: 0;
+        }
+
+        li {
+          border: 1px solid rgba(0, 0, 153, 0.2);
+          width: 50px;
+          height: 50px;
+          border-radius: 100%;
+          margin-right: rem(8px);
+
+          @media screen and (min-width: $screen-m) {
+            margin-right: 0;
+            margin-left: rem(15px);
+          }
+
+          &:hover {
+            background-color: $white;
+          }
+
+          a {
+            width: 100%;
+            height: 100%;
+            display: flex;
+            justify-content: center;
+            align-items: center;
+          }
+        }
+      }
+
+      .footer-contact-us {
+        display: flex;
+        margin-top: rem(20px);
+
+        @media screen and (min-width: $screen-m) {
+          justify-content: flex-end;
+          margin-top: rem(30px);
+        }
+
+        button {
+          font-size: rem(12px);
+          color: #aaa;
+
+          &:hover {
+            color: darken(#aaa, 20%);
+          }
+        }
+      }
+    }
+
+    .footer-partners {
+      padding: rem(30px) 0 0;
+
+      .footer-wri {
+        display: flex;
+        align-items: center;
+        flex-direction: column;
+
+        @media screen and (min-width: $screen-m) {
+          align-content: flex-start;
+        }
+
+        .wri-logo {
+          margin: rem(30px) 0;
+          width: 143px;
+          height: 47px;
+        }
+      }
+
+      .footer-partners-slide {
+        p {
+          text-align: center;
+
+          @media screen and (min-width: $screen-m) {
+            text-align: left;
+          }
+        }
+
+        .slick-list {
+          overflow: hidden;
+        }
+
+        .slick-slide {
+          align-items: center;
+          justify-content: center;
+
+          img {
+            color: black;
+            text-decoration: none;
+            margin: auto;
+          }
+        }
+
+        .slick-arrow {
+          top: calc(50% - 20px);
+        }
+      }
+    }
+  }
+
+  .footer-terms {
+    padding-bottom: rem(25px);
+    color: #999;
+    font-size: 12px;
+    letter-spacing: 0.1px;
+    line-height: 18px;
+    text-align: center;
+
+    @media screen and (min-width: $screen-m) {
+      text-align: right;
+    }
+
+    .terms {
+      display: inline-block;
+      color: inherit;
+      text-decoration: none;
+    }
+
+    .terms:hover {
+      text-decoration: underline;
+    }
+  }
+}
